分离出配置文件；可选hosts源

+++ updateHosts.py
@@ -4,7 +4,7 @@
 #name:updateHosts
 #author:https://github.com/ladder1984
 #python version:2.7.8
-#version:0.0.5
+#version:0.1.0
 ############################
 
 
@@ -13,52 +13,72 @@ from datetime import *
 import re
 import os
 import shutil
+import ConfigParser
 
-#setting
-hostsSource = "https://raw.githubusercontent.com/zxdrive/imouto.host/master/imouto.host.txt"
+#default setting
+hosts_source = "https://raw.githubusercontent.com/zxdrive/imouto.host/master/imouto.host.txt"
 hosts_folder = "C:\Windows\System32\drivers\etc\\"
-hostsLocation = "C:\Windows\System32\drivers\etc\hosts"
+hosts_location = "C:\Windows\System32\drivers\etc\hosts"
 noAdBlock = 1
-#seting
+#default setting
 
 errorLog = open('errorLog.txt', 'a')
 
 
+def get_config():
+    global hosts_source
+    global noAdBlock
+    if os.path.exists('config.ini'):
+        config = ConfigParser.ConfigParser()
+        config.read('config.ini')
+        source_id = config.get('source_select', 'source_id')
+        hosts_source = config.get('source_select', 'source'+source_id)
+        noAdBlock = config.getint('other', 'noadblock')
+
+
 def backup_hosts():
-    if not os.path.isfile(hosts_folder+'backup_hosts_original_by_updateHosts'):
-        shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_original_by_updateHosts')
-    shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_last_by_updateHosts')
+
+    if not os.path.isfile(hosts_folder + 'backup_hosts_original_by_updateHosts'):
+        if os.path.exists(hosts_folder + 'host'):
+            shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_original_by_updateHosts')
+    if os.path.exists(hosts_folder + 'host'):
+        shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_last_by_updateHosts')
 
 
 def download_hosts():
     try:
-        urlretrieve(hostsSource, "hosts")
+        urlretrieve(hosts_source, "hosts")
     except IOError, e:
         errorLog.write(str(datetime.now())+'\n'+str(e)+'\n\n')
 
 
 def del_adblock():
-    file_to_open = open('hosts', 'r')
-    hosts_content = file_to_open.read()
-    hosts_content = re.sub('#AdBlock START([\s\S]*)#AdBlock END', "#", hosts_content)
-    file_to_open.close()
+    if noAdBlock == 1:
+        file_to_open = open('hosts', 'r')
+        hosts_content = file_to_open.read()
+        hosts_content = re.sub('#AdBlock START([\s\S]*)#AdBlock END', "#", hosts_content)
+        file_to_open.close()
 
-    file_to_open = open('hosts', 'w')
-    file_to_open.write(hosts_content)
-    file_to_open.close()
+        file_to_open = open('hosts', 'w')
+        file_to_open.write(hosts_content)
+        file_to_open.close()
 
 
 def move_hosts():
     try:
-        shutil.move("hosts", hostsLocation)
+        shutil.move("hosts", hosts_location)
     except IOError, e:
         errorLog.write(str(datetime.now())+'\n'+str(e)+'\n\n')
 
 errorLog.close()
 
-#main
-backup_hosts()
-download_hosts()
-if noAdBlock == 1:
+
+def main():
+    get_config()
+    backup_hosts()
+    download_hosts()
     del_adblock()
-move_hosts()
+    move_hosts()
+
+if __name__ == '__main__':
+    main()
